refactoring: removed TM from workers
