+++ src/app/core/tm/tm.types.ts
@@ -1,0 +1,10 @@
+export type TMIndexes = {
+  sectionIndex: number;
+  pIndex: number;
+}
+
+export interface TMOptions {
+  bookContent: string,
+  maxChar?: number,
+  indexes?: TMIndexes
+}
